auth provider and dynamic firebase owner!
